server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.1.4
matplotlib==3.8.2
fpdf==1.7.2
google-generativeai==0.3.2
openpyxl==3.1.2
pymongo==4.6.1
python-dotenv==1.0.0 